Move @RenameTo to discord package
