+++ core/src/main/java/me/mastercapexd/auth/discord/command/annotation/RenameTo.java
@@ -1,16 +1,17 @@
-package me.mastercapexd.auth.messenger.commands.annotation;
+package me.mastercapexd.auth.discord.command.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
 /**
  * Required and used in discord module. Used when we need to change {@link revxrsal.commands.annotation.Flag} or {@link revxrsal.commands.annotation.Named}
  * name to another.
- */ public @interface RenameTo {
+ */
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface RenameTo {
     String value();
 
     String type();
